01_urlname 설정&02_views.py 변경(HTTP method)

+++ boards/views.py
@@ -6,31 +6,48 @@ def index(request):
     return render(request, 'boards/index.html', {'board':board})
     
 def new(request):
-    return render(request, 'boards/new.html')
+    if request.method == 'POST':
+        board = boards()
+        board.title = request.POST.get('title')
+        board.content = request.POST.get('content')
+        board.save()
+        return redirect('boards:detail', board.pk)
+    else:
+        return render(request, 'boards/new.html')
     
-def create(request):
-    board = boards()
-    board.title = request.POST.get('title')
-    board.content = request.POST.get('content')
-    board.save()
-    return redirect('/boards/')
+# def create(request):
+#     board = boards()
+#     board.title = request.POST.get('title')
+#     board.content = request.POST.get('content')
+#     board.save()
+#     return redirect('boards:detail', board.pk)
     
 def detail(request, pk):
     board = boards.objects.get(pk=pk)
     return render(request, 'boards/detail.html', {'board':board})
     
 def edit(request, pk):
-    board = boards.objects.get(pk=pk)
-    return render(request, 'boards/edit.html', {'board':board})
+    if request.method == 'POST':
+        board = boards.objects.get(pk=pk)
+        board.title = request.POST.get('title')
+        board.content = request.POST.get('content')
+        board.save()
+        return redirect('boards:detail', board.pk)
+    else:
+        board = boards.objects.get(pk=pk)
+        return render(request, 'boards/edit.html', {'board':board})
     
-def update(request, pk):
-    board = boards.objects.get(pk=pk)
-    board.title = request.POST.get('title')
-    board.content = request.POST.get('content')
-    board.save()
-    return redirect('/boards/')
+# def update(request, pk):
+#     board = boards.objects.get(pk=pk)
+#     board.title = request.POST.get('title')
+#     board.content = request.POST.get('content')
+#     board.save()
+#     return redirect('boards:detail', board.pk)
     
 def delete(request, pk):
     board = boards.objects.get(pk=pk)
-    board.delete()
-    return redirect('/boards/')
+    if request.method == 'POST':
+        board.delete()
+        return redirect('boards:index')
+    else:
+        return redirect('boards:detail', board.pk)
